Update README.asciidoc
Gentoo has a wiki page documenting Kakoune installation. The method given here ("emerge kakoune") will not work on a default Gentoo installation, as Kakoune is currently in the Testing rather than in the Stable branch, so a link to the documentation seems pertinent. The wiki page currently has little content, but does clearly explain how to install Kakoune on Gentoo.